requirements.txt
pygetwindow
pyautogui
Pillow
opencv-python
numpy